Don't explicitly use `as u32` for error enums
`post_error()` takes `impl Into<u32>`.

+++ src/wayland/dmabuf/mod.rs
@@ -1161,7 +1161,7 @@ impl DmabufParamsData {
 
             None => {
                 params.post_error(
-                    zwp_linux_buffer_params_v1::Error::Incomplete as u32,
+                    zwp_linux_buffer_params_v1::Error::Incomplete,
                     "Provided buffer is incomplete, it has zero planes",
                 );
                 return None;
